chore: more clang-tidy cleanups
- explicitly initialize class values
- use size_t instead of ints for sizes
- turn off linting for DDecQuad ctors allowing implicit conversions --
  this is something I want to allow to make usage more natural

+++ src/Boxes.cpp
@@ -216,7 +216,7 @@ Boxes::Box Boxes::FindNextBox (const DprDecimal::DDecQuad& current_value)
         }
     }
 
-    int32_t box_index = ranges::distance(boxes_.begin(), found_it);
+    size_t box_index = ranges::distance(boxes_.begin(), found_it);
     return boxes_.at(box_index + 1);
 }		// -----  end of method Boxes::FindNextBox  -----
